+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - PR #240 Remove deprecated RMM header references.
 - PR #243 Install dependencies via meta package
 - PR #247 Use rmm::device_uvector and cudf::UINT types for quadtree construction
+- PR #246 Hausdorff performance improvement
 
 ## Bug Fixes
 - PR #244 Restrict gdal version